pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Hello World"
license = "GNU Affero"
authors = ["Nicholas M. Synovic <nicholas.synovic@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/NicholasSynovic/template_python"
repository = "https://github.com/NicholasSynovic/template_python"
keywords = ["Template"]
classifiers = ["Development Status :: 1 - Planning"]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.44.2"
streamlit = "^1.38.0"
pillow = "^10.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
torchaudio = "^2.4.0"
sentencepiece = "^0.2.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
